requirements.txt
asgiref==3.8.1
asttokens==2.4.1
Babel==1.3
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
decorator==5.1.1
Django==4.2
django-crispy-forms==2.1
django-extensions==3.2.3
executing==2.0.1
ipython==8.23.0
jedi==0.19.1
jsonfield==1.0.3
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
wcwidth==0.2.13
